requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.7
lxml==4.9.3
schedule==1.2.0
python-dotenv==1.0.0
feedparser==6.0.10
newspaper3k==0.2.8
nltk==3.8.1
sumy==0.11.0
pytz==2023.3
telethon==1.34.0
aiohttp==3.9.1
